// src/types/appartmentDetail.d.ts
export default interface AppartmentDetails {
  id: string
  appartment: string
  createdBy: string
  createdAt: string
  groundLevel: string
  isFurnished: boolean
  hasFloors: boolean
  numberOfFloors: number
  hasBedroom: boolean
  numberOfBedrooms: number
  hasLivingroom: boolean
  numberOfLivingrooms: number
  hasKitchen: boolean
  numberOfKitchen: number
  hasBathrooms: boolean
  numberOfBathrooms: number
  hasStorageroom: boolean
  numberOfStoragerooms: number
  hasGarage: boolean
  numberOfGarage:number
  hasGarden: boolean
  numberOfGarden: number
  hasAC: boolean
  numberOfACs: number
  hasPool: boolean
  numberOfPools: number
  hasKeeper: boolean
  totalOfHouseholds: number
}